Use float32 for bias and tau in examples

+++ examples/lazy_attention_layer.py
@@ -82,7 +82,8 @@ class LazyAttentionLayer(nn.Module):
 
         # === Lazy Attention specific parameters ===
         # Learnable head-wise attention biases [H, max_bias_length]
-        self.bias = nn.Parameter(torch.zeros(self.num_heads, self.max_bias_length))
+        # Use float32 to avoid gradient underflow with bfloat16 training
+        self.bias = nn.Parameter(torch.zeros(self.num_heads, self.max_bias_length, dtype=torch.float32))
         # Learnable sparsity parameter [H]
         # Use float32 to avoid gradient underflow with bfloat16 training
         self.tau = nn.Parameter(torch.full((self.num_heads,), -1.0, dtype=torch.float32))
@@ -159,10 +160,10 @@ class LazyAttentionLayer(nn.Module):
             varlen = (cu_seqlens[1:] - cu_seqlens[:-1]).to(torch.int32)
 
         # === Call Lazy Attention Triton kernel ===
-        # Convert tau from float32 to model dtype for kernel computation
+        # Convert float32 parameters to model dtype for kernel computation
         o = lazy_attention_triton(
             q, k, v,
-            bias=self.bias,
+            bias=self.bias.to(q.dtype),
             tau=self.tau.to(q.dtype),
             window_size=self.max_bias_length,
             varlen=varlen
